check for Utiliser type long a la place int

+++ src/com/projet/fibonacci/FibonacciIteratif.java
@@ -13,11 +13,11 @@ public class FibonacciIteratif {
      * @return un tableau contenant la suite de Fibonacci jusqu'au nombre de termes spécifié
      * @throws IllegalArgumentException si n est négatif
      */
-    public static int[] calculerSuiteFibonacci(int n) {
+    public static long[] calculerSuiteFibonacci(int n) {
         if (n < 0) {
             throw new IllegalArgumentException("Le nombre de termes de la suite de Fibonacci ne peut pas être négatif.");
         }
-        int[] fibNombres = new int[n];
+        long[] fibNombres = new long[n];
         if (n > 0) fibNombres[0] = 0;
         if (n > 1) fibNombres[1] = 1;
         for (int i = 2; i < n; i++) {
